requirements.txt
azure-search-documents==11.4.0a20230509004
openai
python-dotenv